lab_02_a: float -> double

+++ antsor/lab_02/arc.h
@@ -6,7 +6,7 @@
 class Arc : public Outline
 {
 public:
-	Arc(Point p = Point(0, 0), float a = 100, float b = 100, int angle = 180);
+	Arc(Point p = Point(0, 0), double a = 100, double b = 100, int angle = 180);
 private:
 	Point line[POINTS_N];
 };
